Map java script
I have created another webpage for project 3. in this webpage i have created a css designed webpage displaying a map of my home town in sicily and location that I liked visiting. I have used java script to crate the map and markers with click functions to display information window to the enduser so they can get a basic idea of what at that location.